refactor: migrate SubstanceDetail to use ViewModel
Update SubstanceDetailScreen to use a ViewModel for state management, add tabs for detailed insights, and include new fields in Compound data models.

+++ app/src/main/java/com/example/domain/model/Variant.kt
@@ -7,7 +7,7 @@ data class Variant(
     val colorHex: String = "#FFFFFF",
     val pricePerUnit: Float? = null,
     val unitLabel: String = "mg",
-    val ratio: Map<String, Float> = emptyMap(),
+    val ratio: Map<String, Double> = emptyMap(),
     val roaDefault: String = "Oral",
     val active: Boolean = true,
     val createdAt: Long = System.currentTimeMillis(),
